+++ packages/jsenv-plugin-globals/src/main.js
@@ -1,1 +1,2 @@
 export { jsenvPluginGlobals } from "./jsenv_plugin_globals.js"
+export { injectGlobals } from "./inject_globals.js"