+++ packages/jsenv-plugin-placeholders/src/main.js
@@ -1,1 +1,2 @@
 export { jsenvPluginPlaceholders } from "./jsenv_plugin_placeholders.js"
+export { replacePlaceholders } from "./replace_placeholders.js"