+++ src/plugins/plugins.js
@@ -8,6 +8,7 @@ import { jsenvPluginInline } from "./inline/jsenv_plugin_inline.js"
 import { jsenvPluginSupervisor } from "./supervisor/jsenv_plugin_supervisor.js"
 import { jsenvPluginCommonJsGlobals } from "./commonjs_globals/jsenv_plugin_commonjs_globals.js"
 import { jsenvPluginImportMetaScenarios } from "./import_meta_scenarios/jsenv_plugin_import_meta_scenarios.js"
+import { jsenvPluginGlobalScenarios } from "./global_scenarios/jsenv_plugin_global_scenarios.js"
 import { jsenvPluginTranspilation } from "./transpilation/jsenv_plugin_transpilation.js"
 import { jsenvPluginNodeRuntime } from "./node_runtime/jsenv_plugin_node_runtime.js"
 // autoreload
@@ -39,6 +40,7 @@ export const getCorePlugins = ({
   clientFilesPruneCallbackList,
   explorer,
   cacheControl,
+  scenarioPlaceholders = true,
   ribbon = true,
 } = {}) => {
   if (explorer === true) {
@@ -89,6 +91,7 @@ export const getCorePlugins = ({
     jsenvPluginUrlVersion(),
     jsenvPluginCommonJsGlobals(),
     jsenvPluginImportMetaScenarios(),
+    ...(scenarioPlaceholders ? [jsenvPluginGlobalScenarios()] : []),
 
     jsenvPluginNodeRuntime({ runtimeCompat }),