+++ tests/dev_and_build/global_scenario/global_scenario_dev.test.mjs
@@ -1,0 +1,32 @@
+import { assert } from "@jsenv/assert"
+
+import { startDevServer } from "@jsenv/core"
+import { executeInChromium } from "@jsenv/core/tests/execute_in_chromium.js"
+
+const test = async (params) => {
+  const devServer = await startDevServer({
+    logLevel: "warn",
+    rootDirectoryUrl: new URL("./client/", import.meta.url),
+    keepProcessAlive: false,
+    ...params,
+  })
+  const { returnValue } = await executeInChromium({
+    url: `${devServer.origin}/main.html`,
+    /* eslint-disable no-undef */
+    pageFunction: () => window.resultPromise,
+    /* eslint-enable no-undef */
+  })
+  return { returnValue, server: devServer }
+}
+
+const { returnValue } = await test()
+const actual = {
+  returnValue,
+}
+const expected = {
+  returnValue: {
+    dev: true,
+    build: false,
+  },
+}
+assert({ actual, expected })