+++ tests/dev_server/autoreload_js_import_css/client/file.js
@@ -31,3 +31,4 @@ if (import.meta.hot) {
     })
   })
 }
+window.resolveReadyPromise()