+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 31.1.2
+
+- Code can use `__DEV__` and `__BUILD__` to know if code is executed during dev or after build
+  This is complementary to `import.meta.dev/build` for js_classic and workers
+- Fix urls not versioned when minification + vendors chunk is used
+- Fix crash on failed test execution
+- Params check on startBuildServer
+
 # 31.0.0
 
 - Major update on dependencies